+++ src/Collection.php
@@ -4,6 +4,7 @@ namespace DusanKasan\Knapsack;
 
 use Closure;
 use DusanKasan\Knapsack\Exceptions\InvalidArgument;
+use DusanKasan\Knapsack\Exceptions\InvalidReturnValue;
 use Iterator;
 use IteratorAggregate;
 use RecursiveArrayIterator;
@@ -161,5 +162,47 @@ class Collection implements Iterator, \Serializable
     public function unserialize($serialized)
     {
         $this->input = dereferenceKeyValue(unserialize($serialized));
+    }
+
+    /**
+     * Returns a lazy collection of non-lazy collections of items from nth position from this collection and each
+     * passed collection. Stops when any of the collections don't have an item at the nth position.
+     *
+     * @param array|Traversable[] ...$collections
+     * @return Collection
+     */
+    public function zip(...$collections)
+    {
+        array_unshift($collections, $this);
+        return zip(...$collections);
+    }
+
+    /**
+     * Uses a $transformer callable that takes a Collection and returns Collection on itself.
+     *
+     * @param callable $transformer Collection => Collection
+     * @return mixed
+     */
+    public function transform(callable $transformer)
+    {
+        $transformed = $transformer($this);
+
+        if (!($transformed instanceof Collection)) {
+            throw new InvalidReturnValue;
+        }
+
+        return $transformed;
+    }
+
+    /**
+     * Extracts data from collection items by dot separated key path. Supports the * wildcard.  If a key contains \ or
+     * it must be escaped using \ character.
+     *
+     * @param mixed $keyPath
+     * @return Collection
+     */
+    public function extract($keyPath)
+    {
+        return \DusanKasan\Knapsack\extract($this, $keyPath);
     }
 }