+++ src/Exceptions/InvalidReturnValue.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace DusanKasan\Knapsack\Exceptions;
+
+class InvalidReturnValue extends RuntimeException
+{
+}