+++ src/Exceptions/NonEqualCollectionLength.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace DusanKasan\Knapsack\Exceptions;
-
-class NonEqualCollectionLength extends RuntimeException
-{
-}