+++ tests/spec/Exceptions/InvalidReturnValueSpec.php
@@ -6,10 +6,7 @@ use DusanKasan\Knapsack\Exceptions\RuntimeException;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
-/**
- * @mixin NonEqualCollectionLengthSpec
- */
-class NonEqualCollectionLengthSpec extends ObjectBehavior
+class InvalidReturnValueSpec extends ObjectBehavior
 {
     function it_is_initializable()
     {